Embed: Move SetFrameStrata call to a better place
Figured out what was causing the strata to change/be unreliable.
It's due to the SetParent call, which causes the frame to inherit
the strata of its new parent.

Knowing this, we can now simply set the strata *after* SetParent,
rather than hooking the whole Show() function! A bit cleaner!