Clique: Fix ugly checkbox backdrop bug
The skin was setting up a nice checkbox design, but then Clique
itself constantly modified the textures again whenever the lists
were updated (in "Clique:TextListScrollUpdate()").

As a result, all checkboxes looked ugly as hell, with stretched,
weird texture artifacts.

To fix that problem, we apply our skin and then we turn all "change
texture" functions into no-ops, so that Clique can't mess with it.