TinyBook: Add support for TinyBook v1.2.0
Adds support for the "TinyBook" spellbook replacement addon.

This is a fresh, clean and compact spellbook which only shows the
highest rank of each spell. And if you ever want to downrank your
spells to a lower rank, you can simply hover over a spell to see
all other ranks. Easy and intuitive!

https://github.com/VideoPlayerCode/TinyBook

(This new ElvUI skin perfectly matches the regular ElvUI spellbook.)